refactor(pool-page): extract _tabs/ols-tab
Move OlsTab (~45 lines) out of page.tsx into _tabs/ols-tab.tsx. Tab
imports from _components/ for OlsStatusPanel + OlsLiquidityEvents and
from _lib/helpers for selectActiveOlsPool — no import cycle with
page.tsx.

Validates the per-tab extraction pattern on the smallest tab; subsequent
commits will follow the same shape for the remaining six tabs.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=ols vs baseline:
  zero structural changes, console clean

page.tsx: 2211 → 2169 lines (-42).
